first pass at functional Create Spot form
